refactor: move extension capability structs to homeboy-extension-contract
Eleventh sub-cut of the extension-contract campaign (Stage 1, wiki 12840).

With all their field sub-trees now in contract crates (from earlier cuts),
the five pure capability structs become movable: DeployCapability,
AuditCapability, ExecutableCapability, PlatformCapability, AgentTaskPolicyConfig.

Moves them into a new contract manifest_capabilities module. Their field types
resolve from the extension-contract submodules (deploy/toolchain/action/capability
config) plus homeboy-audit-contract (AuditConfig, TestMappingConfig). All five
are impl-free with no custom serde helpers.

manifest.rs re-exports them (pub use), keeping the crate::extension::* surface
stable. This leaves manifest.rs with just the behavior-bearing types
(ExtensionManifest ConfigEntity, ExtensionDiagnosticsConfig,
NotificationTransportConfig) - the Stage 2 boundary.
